refactor: ♻️ extract pull_issues_from_github to pull.rs (#191)
